Now loads data & WP core functionality
In integration.php I need to use an absolute windows path. This is because the folder in the plugins directory of the server is a symlink to centris-integration-plugin, and the server runs on a windows based shell. In the future, the path to wp-load.php will be in relative referencing for production environments.